ana sayfa düzenlendi

+++ app/src/main/java/com/yarenbergi/yemekbul/data/Metric.kt
@@ -3,7 +3,7 @@ package com.yarenbergi.yemekbul.data
 import com.google.gson.annotations.SerializedName
 
 data class Metric(@SerializedName("amount")
-                  val amount: Int = 0,
+                  val amount: Double = 0.0,
                   @SerializedName("unitShort")
                   val unitShort: String = "",
                   @SerializedName("unitLong")
